+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/extensions/extensions.kt
@@ -4,13 +4,8 @@
 
 package org.jetbrains.intellij.platform.gradle.extensions
 
-import kotlinx.serialization.encodeToString
-import nl.adaptivity.xmlutil.serialization.XML
 import org.gradle.api.Project
 import org.gradle.api.artifacts.Dependency
-import org.gradle.api.artifacts.ExternalModuleDependency
-import org.gradle.api.artifacts.dsl.RepositoryHandler
-import org.gradle.api.artifacts.repositories.UrlArtifactRepository
 import org.gradle.api.file.Directory
 import org.gradle.api.plugins.ExtensionContainer
 import org.gradle.api.provider.ProviderFactory
@@ -18,72 +13,11 @@ import org.gradle.internal.os.OperatingSystem
 import org.jetbrains.intellij.platform.gradle.Constants.CACHE_DIRECTORY
 import org.jetbrains.intellij.platform.gradle.Constants.GradleProperties
 import org.jetbrains.intellij.platform.gradle.IntelliJPlatformType
-import org.jetbrains.intellij.platform.gradle.models.IvyModule
 import org.jetbrains.intellij.platform.gradle.toIntelliJPlatformType
 import org.jetbrains.intellij.platform.gradle.utils.asPath
 import java.io.File
 import java.nio.file.Path
 import kotlin.io.path.*
-
-/**
- * Creates an Ivy dependency XML file for an external module.
- *
- * @param localPlatformArtifactsPath The [Path] to the local IntelliJ Platform artifacts directory.
- * @param publications The list of [IvyModule.Publication] objects to be included in the Ivy file.
- */
-internal fun ExternalModuleDependency.createIvyDependencyFile(
-    localPlatformArtifactsPath: Path,
-    publications: List<IvyModule.Artifact>,
-    dependencies: List<IvyModule.Dependency> = emptyList(),
-) =
-    createIvyDependencyFile(
-        group = group,
-        name = name,
-        version = version.orEmpty(),
-        localPlatformArtifactsPath = localPlatformArtifactsPath,
-        publications = publications,
-        dependencies = dependencies,
-    )
-
-/**
- * Creates an Ivy dependency XML file for an external module.
- *
- * @param localPlatformArtifactsPath The [Path] to the local IntelliJ Platform artifacts directory.
- * @param publications The list of [IvyModule.Publication] objects to be included in the Ivy file.
- */
-internal fun createIvyDependencyFile(
-    group: String?,
-    name: String,
-    version: String,
-    localPlatformArtifactsPath: Path,
-    publications: List<IvyModule.Artifact>,
-    dependencies: List<IvyModule.Dependency> = emptyList(),
-) {
-    val ivyFile = localPlatformArtifactsPath
-        .resolve("$group-$name-$version.xml")
-// TODO: ignore existing Ivy files as path may be invalid over time thanks to Gradle transformers rotation
-//        .takeUnless { it.exists() }
-//        ?: return
-
-    val ivyModule = IvyModule(
-        info = IvyModule.Info(
-            organisation = group,
-            module = name,
-            revision = version,
-        ),
-        publications = publications,
-        dependencies = dependencies
-    )
-
-    with(ivyFile) {
-        parent.createDirectories()
-        deleteIfExists()
-        createFile()
-        writeText(XML {
-            indentString = "  "
-        }.encodeToString(ivyModule))
-    }
-}
 
 @Throws(IllegalArgumentException::class)
 internal fun resolvePath(localPath: Any) = when (localPath) {
@@ -135,17 +69,12 @@ internal fun ProviderFactory.localPlatformArtifactsPath(rootProjectDirectory: Pa
 /**
  * Retrieves URLs from registered repositories.
  */
-internal fun RepositoryHandler.urls() = mapNotNull { (it as? UrlArtifactRepository)?.url?.toString() }
 internal fun String.parseIdeNotation() = trim().split('-').let {
     when {
         it.size == 2 -> it.first().toIntelliJPlatformType() to it.last()
         else -> IntelliJPlatformType.IntellijIdeaCommunity to it.first()
     }
 }
-/**
- * Type alias for a lambda function that takes a [Dependency] and performs some actions on it.
- */
-internal typealias DependencyAction = (Dependency.() -> Unit)
 
 /**
  * Parses the plugin notation into the `<id, version, channel` triple.
@@ -162,6 +91,10 @@ internal fun String.parsePluginNotation() = trim()
  * Because [ExtensionContainer.create] accepts extension arguments provided with no strong typing,
  */
 internal interface Registrable<T> {
-
     fun register(project: Project, target: Any): T
 }
+
+/**
+ * Type alias for a lambda function that takes a [Dependency] and performs some actions on it.
+ */
+internal typealias DependencyAction = (Dependency.() -> Unit)